Use %28 and %29 instead of parentheses in MarkDown
While ( and ) are legal characters in URLs, they don't mix particularly
well with the MarkDown syntax.

Annoyingly this makes the pages look a bit more incoherent in their
source.  Which is a bit of a strike against MarkDown's mixture with
the web.

But oh well, here's a start to doing some updates circa April 2018.

Not dead.  (Yet.)